new file:   Fullstack/resources/js/Pages/Listing/Show.vue

+++ Fullstack/database/migrations/2023_02_08_040503_add_fields_to_listings_table.php
@@ -16,14 +16,14 @@ return new class extends Migration
         Schema::table('listings', function (Blueprint $table) {
             $table->unsignedTinyInteger('beds');
             $table->unsignedTinyInteger('baths');
-            $table->unsignedTinyInteger('area');
+            $table->unsignedSmallInteger('area');
 
             $table->tinyText('city');
             $table->tinyText('code');
             $table->tinyText('street');
             $table->tinyText('street_nr');
 
-            $table->unsignedSmallInteger('price');
+            $table->unsignedInteger('price');
         });
     }
